Implement classify in C++
I still need a make_tree wrapper. Also, I'm deviating slightly from the
morloc implementation by keeping the int index in the tree and an
external table of clades and annotations.

+++ flu-case-comparisons/modules/rooted_tree/src/kmer.hpp
@@ -2,14 +2,16 @@
 #define __KMER_HPP__
 
 // adapted from morloc bio.algo module
-std::map<std::string,int> countKmers(int k, std::string seq){
+std::map<std::string,int>
+countKmers(int k, std::string seq)
+{
     std::map<std::string,int> kmers;
    
     if(k <= 0){
         throw std::invalid_argument("k must be an integer greater than or equal to 1");
     }
 
-    for(int i = 0; (i + k) <= seq.size(); i++){
+    for(size_t i = 0; (i + k) <= seq.size(); i++){
         std::string kmer = seq.substr(i, k);
         if(kmers.find(kmer) != kmers.end()){
             kmers[kmer]++;
@@ -24,7 +26,9 @@ std::map<std::string,int> countKmers(int k, std::string seq){
 
 
 // adapted from morloc bio.algo module
-double kmerDistance(const std::map<std::string,int>& x, const std::map<std::string,int>& y){
+double
+kmerDistance(const std::map<std::string,int>& x, const std::map<std::string,int>& y)
+{
     double square_distance = 0.0;
 
     // Iterate through kmers in sequence x, find distance to y
@@ -56,7 +60,9 @@ double kmerDistance(const std::map<std::string,int>& x, const std::map<std::stri
 
 // adapted from morloc matrix module
 template <typename A, typename B>
-Eigen::Matrix<B, Eigen::Dynamic, Eigen::Dynamic> selfcmp(std::function<B(A, A)> f, std::vector<A> xs){
+Eigen::Matrix<B, Eigen::Dynamic, Eigen::Dynamic>
+selfcmp(std::function<B(A, A)> f, std::vector<A> xs)
+{
     int size = xs.size();
 
     Eigen::Matrix<B, Eigen::Dynamic, Eigen::Dynamic> mat(size, size);
@@ -72,7 +78,8 @@ Eigen::Matrix<B, Eigen::Dynamic, Eigen::Dynamic> selfcmp(std::function<B(A, A)> 
 
 
 Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>
-makeDistMat(int kmer_size, std::vector<std::string> seqs){
+makeDistMat(int kmer_size, std::vector<std::string> seqs)
+{
   std::vector<std::map<std::string,int>> kmermaps;
   for (size_t i = 0; i < seqs.size(); i++){
     std::map<std::string,int> kmermap = countKmers(kmer_size, seqs[i]);
